Move interface into interfaces subdirectory
Also make naming consistent with other parts of the application.
