+++ BonusCalcApi.Tests/V1/UseCase/GetPayElementTypeUseCaseTests.cs
@@ -2,6 +2,7 @@ using System.Threading.Tasks;
 using AutoFixture;
 using BonusCalcApi.Tests.V1.Helpers;
 using BonusCalcApi.V1.Gateways;
+using BonusCalcApi.V1.Gateways.Interfaces;
 using BonusCalcApi.V1.Infrastructure;
 using BonusCalcApi.V1.UseCase;
 using FluentAssertions;
