+++ BonusCalcApi/V1/Gateways/Interfaces/IPayElementTypesGateway.cs
@@ -2,7 +2,7 @@ using BonusCalcApi.V1.Infrastructure;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace BonusCalcApi.V1.Gateways
+namespace BonusCalcApi.V1.Gateways.Interfaces
 {
     public interface IPayElementTypesGateway
     {
