+++ BonusCalcApi/V1/Gateways/PayElementTypeGateway.cs
@@ -1,9 +1,10 @@
-using BonusCalcApi.V1.Infrastructure;
-using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.EntityFrameworkCore;
+using BonusCalcApi.V1.Gateways.Interfaces;
+using BonusCalcApi.V1.Infrastructure;
 
 namespace BonusCalcApi.V1.Gateways
 {
