+++ BonusCalcApi/V1/UseCase/GetPayElementTypeUseCase.cs
@@ -1,25 +1,24 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
 using BonusCalcApi.V1.Gateways;
+using BonusCalcApi.V1.Gateways.Interfaces;
 using BonusCalcApi.V1.Infrastructure;
 using BonusCalcApi.V1.UseCase.Interfaces;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace BonusCalcApi.V1.UseCase
 {
     public class GetPayElementTypeUseCase : IGetPayElementTypeUseCase
     {
-        private IPayElementTypesGateway _payElementType;
+        private IPayElementTypesGateway _payElementTypesGateway;
 
-        public GetPayElementTypeUseCase(IPayElementTypesGateway payElementType)
+        public GetPayElementTypeUseCase(IPayElementTypesGateway payElementTypesGateway)
         {
-            _payElementType = payElementType;
+            _payElementTypesGateway = payElementTypesGateway;
         }
 
         public async Task<IEnumerable<PayElementType>> ExecuteAsync()
         {
-            return await _payElementType.GetPayElementTypesAsync();
+            return await _payElementTypesGateway.GetPayElementTypesAsync();
         }
     }
 }
